+++ client/state/guided-tours/contexts/has-selected-site-defaults-site-title.js
@@ -1,0 +1,16 @@
+/**
+ * Internal dependencies
+ */
+import { getSelectedSiteId } from 'state/ui/selectors';
+import { hasDefaultSiteTitle } from 'state/sites/selectors';
+
+/**
+ * Returns true if the selected site has an unchanged site title
+ *
+ * @param {object} state Global state tree
+ * @returns {boolean} True if site title is default, false otherwise.
+ */
+export const hasSelectedSiteDefaultSiteTitle = ( state ) => {
+	const siteId = getSelectedSiteId( state );
+	return siteId ? hasDefaultSiteTitle( state, siteId ) : false;
+};
