+++ client/state/guided-tours/contexts/has-user-registered-before.js
@@ -1,0 +1,17 @@
+/**
+ * Internal dependencies
+ */
+import { getCurrentUser } from 'state/current-user/selectors';
+
+/**
+ * Returns a selector that tests if the user has registered before given date
+ *
+ * @param {Date} date Date of registration
+ * @returns {Function} Selector function
+ */
+export const hasUserRegisteredBefore = ( date ) => ( state ) => {
+	const compareDate = date && Date.parse( date );
+	const user = getCurrentUser( state );
+	const registrationDate = user && Date.parse( user.date );
+	return registrationDate < compareDate;
+};
