+++ client/state/guided-tours/contexts/in-section.js
@@ -1,0 +1,12 @@
+/**
+ * Internal dependencies
+ */
+import { getSectionName } from 'state/ui/selectors';
+
+/**
+ * Returns a selector that tests if the current user is in a given section
+ *
+ * @param {string} sectionName Name of section
+ * @returns {Function} Selector function
+ */
+export const inSection = ( sectionName ) => ( state ) => getSectionName( state ) === sectionName;
