+++ client/state/guided-tours/contexts/is-enabled.js
@@ -1,0 +1,12 @@
+/**
+ * Internal dependencies
+ */
+import config from 'config';
+
+/**
+ * Returns a selector that tests if a feature is enabled in config
+ *
+ * @param {string} feature Name of feature
+ * @returns {Function} Selector function
+ */
+export const isEnabled = ( feature ) => () => config.isEnabled( feature );
