+++ client/state/guided-tours/contexts/is-selected-site-customizable.js
@@ -1,0 +1,13 @@
+/**
+ * Internal dependencies
+ */
+import { getSelectedSite } from 'state/ui/selectors';
+
+/**
+ * Returns true if the current user can run customizer for the selected site
+ *
+ * @param {object} state Global state tree
+ * @returns {boolean} True if user can run customizer, false otherwise.
+ */
+export const isSelectedSiteCustomizable = ( state ) =>
+	getSelectedSite( state ) && getSelectedSite( state ).is_customizable;
